feat(email): merge "custom-mailjet-default-sender" endpoint into "mailjet-configuration" endpoint
