+++ src/v0/email.ts
@@ -11,6 +11,7 @@ export interface EmailOptions {
 export interface MailjetCredentials {
   apiKey: string
   apiSecret: string
+  defaultSenderMail?: string
 }
 
 export interface MailjetEmail {
@@ -49,17 +50,6 @@ export interface TestCustomMailjetResponse {
   status?: number
 }
 
-export interface CustomMailjetDefaultSenderRequest {
-  email: string
-}
-
-export interface CustomMailjetDefaultSenderResponse {
-  data?: {
-    success: boolean
-  }
-  msg?: string
-  status?: number
-}
 
 export interface CustomMailjetCredentialStatusResponse {
   data?: {
@@ -149,24 +139,6 @@ export class Email extends ThBaseHandler {
     }
   }
 
-  async setCustomMailjetDefaultSender (request: CustomMailjetDefaultSenderRequest): Promise<CustomMailjetDefaultSenderResponse> {
-    try {
-      const uri = this.uriHelper.generateBaseUri('/custom-mailjet-default-sender')
-      const response = await this.http.getClient().post(uri, request)
-
-      if (response.status !== 200) {
-        throw new errors.EmailCustomMailjetDefaultSenderSetFailed(undefined, { status: response.status })
-      }
-
-      return {
-        data: response.data.results,
-        msg: response.data.msg,
-        status: response.data.status
-      }
-    } catch (error: any) {
-      throw new errors.EmailCustomMailjetDefaultSenderSetFailed(error.message, { error })
-    }
-  }
 
   async getCustomMailjetCredentialStatus (): Promise<CustomMailjetCredentialStatusResponse> {
     try {
